move setMappings from nex to regulare controller
